# contracts/openapi.yaml
openapi: 3.1.0
info:
  title: ID8 Operator API
  version: 0.1.0
  description: |
    MVP API contract for orchestrating prompt-to-production runs with HITL gates.
servers:
  - url: https://api.id8.local
paths:
...
  /v1/projects/{projectId}/runs:
    post:
      summary: Start or resume a run
      operationId: createRun
      parameters:
        - $ref: '#/components/parameters/ProjectId'
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateRunRequest'
      responses:
        '202':
          description: Run accepted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProjectRun'
  /v1/projects/{projectId}/runs/latest:
    get:
      summary: Get latest run with timeline events
      operationId: getLatestRun
      parameters:
        - $ref: '#/components/parameters/ProjectId'
      responses:
        '200':
          description: Latest run detail
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProjectRunDetail'
...
components:
  parameters:
    ProjectId:
      name: projectId
      in: path
      required: true
      schema:
        type: string
        format: uuid
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: false
      schema:
        type: string
  schemas:
...
    ModelProfile:
      type: string
      enum:
        - primary
        - customtools
        - fallback
    ProjectStatus:
      type: string
      enum:
        - ideation
        - prd_draft
        - prd_approved
        - design_draft
        - design_approved
        - tech_plan_draft
        - tech_plan_approved
        - codegen
        - security_gate
        - deploy_ready
        - deploying
        - deployed
        - failed
...
    CreateRunRequest:
      type: object
      properties:
        resume_from_node:
          type: string
        model_profile:
          $ref: '#/components/schemas/ModelProfile'
...
    ProjectRun:
      type: object
      properties:
        id:
          type: string
          format: uuid
        project_id:
          type: string
          format: uuid
        status:
          $ref: '#/components/schemas/ProjectStatus'
        current_node:
          type: string
        idempotency_key:
          type: string
        last_error_code:
          type: string
        last_error_message:
          type: string
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
      required:
        - id
        - project_id
        - status
        - current_node
        - created_at
        - updated_at
    RunTimelineEvent:
      type: object
      properties:
        event_type:
          type: string
        from_node:
          type: string
        to_node:
          type: string
        outcome:
          type: string
        created_at:
          type: string
          format: date-time
      required:
        - event_type
        - to_node
        - created_at
    ProjectRunDetail:
      allOf:
        - $ref: '#/components/schemas/ProjectRun'
        - type: object
          properties:
            timeline:
              type: array
              items:
                $ref: '#/components/schemas/RunTimelineEvent'
          required:
            - timeline
...